# backend/app/api/routers/lookups.py
from fastapi import APIRouter, Depends, Query
from sqlalchemy import select
from sqlalchemy.orm import Session

from app.core.auth import get_current_user
from app.core.database import get_db
from app.core.permissions import ensure_update, permissions_for, require
from app.models.team import Team
from app.models.user import User
from app.schemas.lookup import (
    CategoryCreateRequest,
    CategoryOut,
    CategoryUpdateRequest,
    SLASettingsOut,
    SLASettingsUpdateRequest,
    TeamOut,
)
from app.schemas.user import MeOut, SettingsPermissions, UserOut
from app.services import lookup_service, sla_service

router = APIRouter(tags=["lookups"])


@router.get("/teams", response_model=list[TeamOut])
def list_teams(db: Session = Depends(get_db)) -> list[TeamOut]:
    return db.execute(select(Team).order_by(Team.name)).scalars().all()


@router.get("/categories", response_model=list[CategoryOut])
def list_categories(include_archived: bool = Query(default=False), db: Session = Depends(get_db)) -> list[CategoryOut]:
    return lookup_service.list_categories(db, include_archived=include_archived)


@router.post("/categories", response_model=CategoryOut, status_code=201)
def create_category(
    payload: CategoryCreateRequest, db: Session = Depends(get_db), _: User = Depends(require("create"))
) -> CategoryOut:
    return lookup_service.create_category(db, payload.name)


@router.patch("/categories/{category_id}", response_model=CategoryOut)
def update_category(
    category_id: int, payload: CategoryUpdateRequest, db: Session = Depends(get_db), user: User = Depends(get_current_user)
) -> CategoryOut:
    changes = payload.model_dump(exclude_unset=True)
    ensure_update(db, user, changes)
    return lookup_service.update_category(db, category_id, **changes)


@router.get("/sla-settings", response_model=SLASettingsOut)
def get_sla_settings(db: Session = Depends(get_db)) -> SLASettingsOut:
    return sla_service.get_settings(db)


@router.patch("/sla-settings", response_model=SLASettingsOut)
def update_sla_settings(
    payload: SLASettingsUpdateRequest, db: Session = Depends(get_db), _: User = Depends(require("edit"))
) -> SLASettingsOut:
    return sla_service.set_default_hours(db, payload.default_hours)


@router.get("/me", response_model=MeOut)
def me(db: Session = Depends(get_db), user: User = Depends(get_current_user)) -> MeOut:
    return MeOut(user=UserOut.model_validate(user), settings=SettingsPermissions(**permissions_for(db, user)))


@router.get("/users", response_model=list[UserOut])
def list_users(db: Session = Depends(get_db)) -> list[UserOut]:
    return db.execute(select(User).where(User.is_active.is_(True)).order_by(User.name)).scalars().all()