# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CategoryCreateRequest": {
        "properties": {
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CategoryCreateRequest",
        "type": "object"
      },
      "CategoryOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_archived": {
            "default": false,
            "title": "Is Archived",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "CategoryOut",
        "type": "object"
      },
      "CategoryUpdateRequest": {
        "properties": {
          "is_archived": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Archived"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "CategoryUpdateRequest",
        "type": "object"
      },
      "CustomFieldDefinitionCreateRequest": {
        "properties": {
          "field_type": {
            "$ref": "#/components/schemas/CustomFieldType"
          },
          "label": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          }
        },
        "required": [
          "label",
          "field_type"
        ],
        "title": "CustomFieldDefinitionCreateRequest",
        "type": "object"
      },
      "CustomFieldDefinitionOut": {
        "properties": {
          "field_type": {
            "$ref": "#/components/schemas/CustomFieldType"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_archived": {
            "title": "Is Archived",
            "type": "boolean"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          }
        },
        "required": [
          "id",
          "label",
          "field_type",
          "options",
          "is_archived"
        ],
        "title": "CustomFieldDefinitionOut",
        "type": "object"
      },
      "CustomFieldDefinitionUpdateRequest": {
        "properties": {
          "is_archived": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Archived"
          },
          "label": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          }
        },
        "title": "CustomFieldDefinitionUpdateRequest",
        "type": "object"
      },
      "CustomFieldType": {
        "enum": [
          "text",
          "dropdown"
        ],
        "title": "CustomFieldType",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MeOut": {
        "description": "The current user and what they may do in Settings (see app.core.permissions).",
        "properties": {
          "settings": {
            "$ref": "#/components/schemas/SettingsPermissions"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "required": [
          "user",
          "settings"
        ],
        "title": "MeOut",
        "type": "object"
      },
...
      "RoleOut": {
        "properties": {
          "can_create_settings": {
            "title": "Can Create Settings",
            "type": "boolean"
          },
          "can_delete_settings": {
            "title": "Can Delete Settings",
            "type": "boolean"
          },
          "can_edit_settings": {
            "title": "Can Edit Settings",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_archived": {
            "title": "Is Archived",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "can_create_settings",
          "can_edit_settings",
          "can_delete_settings",
          "is_archived"
        ],
        "title": "RoleOut",
        "type": "object"
      },
...
      "SLASettingsOut": {
        "properties": {
          "default_hours": {
            "title": "Default Hours",
            "type": "integer"
          }
        },
        "required": [
          "default_hours"
        ],
        "title": "SLASettingsOut",
        "type": "object"
      },
      "SLASettingsUpdateRequest": {
        "properties": {
          "default_hours": {
            "exclusiveMinimum": 0.0,
            "title": "Default Hours",
            "type": "integer"
          }
        },
        "required": [
          "default_hours"
        ],
        "title": "SLASettingsUpdateRequest",
        "type": "object"
      },
      "SettingsPermissions": {
        "properties": {
          "create": {
            "title": "Create",
            "type": "boolean"
          },
          "delete": {
            "title": "Delete",
            "type": "boolean"
          },
          "edit": {
            "title": "Edit",
            "type": "boolean"
          }
        },
        "required": [
          "create",
          "edit",
          "delete"
        ],
        "title": "SettingsPermissions",
        "type": "object"
      },
...
      "TeamCreateRequest": {
        "properties": {
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "TeamCreateRequest",
        "type": "object"
      },
      "TeamDetailOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_default": {
            "default": false,
            "title": "Is Default",
            "type": "boolean"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/TeamMemberOut"
            },
            "title": "Members",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "members"
        ],
        "title": "TeamDetailOut",
        "type": "object"
      },
      "TeamMemberOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "role": {
            "$ref": "#/components/schemas/RoleOut"
          },
          "user": {
            "$ref": "#/components/schemas/UserOut"
          }
        },
        "required": [
          "id",
          "user",
          "role"
        ],
        "title": "TeamMemberOut",
        "type": "object"
      },
      "TeamOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_default": {
            "default": false,
            "title": "Is Default",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "TeamOut",
        "type": "object"
      },
...
      "UserOut": {
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slack_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slack User Id"
          }
        },
        "required": [
          "id",
          "email",
          "name"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Support Operations Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/categories": {
      "get": {
        "operationId": "list_categories_api_categories_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoryOut"
                  },
                  "title": "Response List Categories Api Categories Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Categories",
        "tags": [
          "lookups"
        ]
      },
      "post": {
        "operationId": "create_category_api_categories_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-dev-user-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Category",
        "tags": [
          "lookups"
        ]
      }
    },
    "/api/categories/{category_id}": {
      "patch": {
        "operationId": "update_category_api_categories__category_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "category_id",
            "required": true,
            "schema": {
              "title": "Category Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Category",
        "tags": [
          "lookups"
        ]
      }
    },
    "/api/custom-fields": {
      "get": {
        "operationId": "list_custom_fields_api_custom_fields_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CustomFieldDefinitionOut"
                  },
                  "title": "Response List Custom Fields Api Custom Fields Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Custom Fields",
        "tags": [
          "custom-fields"
        ]
      },
      "post": {
        "operationId": "create_custom_field_api_custom_fields_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-dev-user-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomFieldDefinitionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomFieldDefinitionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Custom Field",
        "tags": [
          "custom-fields"
        ]
      }
    },
    "/api/custom-fields/{field_id}": {
      "patch": {
        "operationId": "update_custom_field_api_custom_fields__field_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "field_id",
            "required": true,
            "schema": {
              "title": "Field Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "x-dev-user-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomFieldDefinitionUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomFieldDefinitionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Custom Field",
        "tags": [
          "custom-fields"
        ]
      }
    },
    "/api/me": {
      "get": {
        "operationId": "me_api_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-dev-user-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Me",
        "tags": [
          "lookups"
        ]
      }
    },
...
    "/api/sla-settings": {
      "get": {
        "operationId": "get_sla_settings_api_sla_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SLASettingsOut"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Sla Settings",
        "tags": [
          "lookups"
        ]
      },
      "patch": {
        "operationId": "update_sla_settings_api_sla_settings_patch",
        "parameters": [
          {
            "in": "header",
            "name": "x-dev-user-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SLASettingsUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SLASettingsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Sla Settings",
        "tags": [
          "lookups"
        ]
      }
    },
...
    "/api/teams": {
      "get": {
        "operationId": "list_teams_api_teams_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TeamOut"
                  },
                  "title": "Response List Teams Api Teams Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Teams",
        "tags": [
          "lookups"
        ]
      },
      "post": {
        "operationId": "create_team_api_teams_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-dev-user-email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dev-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Team",
        "tags": [
          "teams"
        ]
      }
    },
...
  }
}